+++ src/test/java/security/validators/TempfileMktempInspectionTest.kt
@@ -1,17 +1,17 @@
 package security.validators
 
+import com.intellij.codeInspection.LocalInspectionToolSession
+import com.intellij.codeInspection.LocalQuickFix
+import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.lang.annotation.Annotation
-import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiTreeUtil
 import com.jetbrains.python.PythonFileType
+import com.jetbrains.python.inspections.PyInspectionVisitor
 import com.jetbrains.python.psi.PyCallExpression
-import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.doReturn
-import com.nhaarman.mockitokotlin2.eq
-import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.*
 import org.jetbrains.annotations.NotNull
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
@@ -20,9 +20,10 @@ import org.junit.jupiter.api.TestInstance
 import org.mockito.Mockito
 import security.Checks
 import security.SecurityTestTask
+import kotlin.check
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class PyyamlLoadValidatorTest: SecurityTestTask() {
+class TempfileMktempInspectionTest: SecurityTestTask() {
     lateinit var dummyAnnotation: Annotation
 
     @BeforeAll
@@ -37,39 +38,42 @@ class PyyamlLoadValidatorTest: SecurityTestTask() {
     }
 
     @Test
-    fun `test yaml load`(){
+    fun `test temp file with insecure make`(){
         var code = """
-            import yaml
-            yaml.load()
+            import tempfile
+            tempfile.mktemp()
         """.trimIndent()
         testCodeString(code, 1)
     }
 
     @Test
-    fun `test yaml safe_load`(){
+    fun `test temp file with make (safe) temp`(){
         var code = """
-            import yaml
-            yaml.safe_load()
+            import tempfile
+            mkstemp()
         """.trimIndent()
         testCodeString(code, 0)
     }
 
     private fun testCodeString(code: String, times: Int = 1){
-        val mockHolder = mock<AnnotationHolder> {
-            on { createWarningAnnotation(any<PsiElement>(), eq(Checks.PyyamlUnsafeLoadCheck.toString())) } doReturn(dummyAnnotation);
+        val mockHolder = mock<ProblemsHolder> {
+            on { registerProblem(any<PsiElement>(), eq(Checks.TempfileMktempCheck.getDescription()), any<LocalQuickFix>()) } doAnswer {}
         }
         ApplicationManager.getApplication().runReadAction {
             val testFile = this.createLightFile("test.py", PythonFileType.INSTANCE.language, code);
+            val mockLocalSession = mock<LocalInspectionToolSession> {
+                on { file } doReturn (testFile)
+            }
             assertNotNull(testFile)
-            val testValidator = PyyamlLoadValidator()
-            testValidator.holder = mockHolder
+            val testVisitor = TempfileMktempInspection().buildVisitor(mockHolder, true, mockLocalSession) as PyInspectionVisitor
 
             val expr: @NotNull MutableCollection<PyCallExpression> = PsiTreeUtil.findChildrenOfType(testFile, PyCallExpression::class.java)
             assertNotNull(expr)
             expr.forEach { e ->
-                testValidator.visitPyCallExpression(e)
+                testVisitor.visitPyCallExpression(e)
             }
-            Mockito.verify(mockHolder, Mockito.times(times)).createWarningAnnotation(any<PsiElement>(), eq(Checks.PyyamlUnsafeLoadCheck.toString()))
+            Mockito.verify(mockHolder, Mockito.times(times)).registerProblem(any<PsiElement>(), eq(Checks.TempfileMktempCheck.getDescription()), any<LocalQuickFix>())
+            Mockito.verify(mockLocalSession, Mockito.times(1)).file
         }
     }
 }